requirements.txt
google-auth
google-api-python-client
google-auth-httplib2
boto3
beautifulsoup4
